Collection Page UI Details Headache
Updated the UI for the collection page to display additional details.
Initial idea was to display a detail panel every time a singular item is clicked.
When another item is clicked, it closes all open details panel so only one exists.
Was to utilize the TabButton and TabGroup scripts to make new scripts to CollectionUIButton and CollectionUIGroup.
Even made a UI Container for the details panel, which I need to remove later.

Idea failed due to the prefab being unable to reference anything outside of itself.
Thus settle for a simpler idea utilizing the button component, and whatever remains that 'inventory UI' video had with the setItemDetails method.

Which another issue arise with the collectionUIController instance being null, was simple fix, but what a headache to find out what the issue was.

Time wasted just doing this simple update.